Homepage code
Images and Slideshow pictures may need to be put in the proper content/content types/directories.
The homepage still needs a review/comment section. I will implement that by 11/11/15. the client will let us know if they like the page layout
Lastly, please test the page. The login/signup should be working. For now, the login just lets you know whether it was successful or not. Sign up should create a new account if you set up the content type properly and change the stInode # for your local machine. The navbar is not fixed (I will change that). I am also working on a footer. The links are there for show, but once we have one master depository* they should work.